requirements.txt
Flask
Flask-SQLAlchemy
Authlib==0.14.3
paho-mqtt
Flask-BasicAuth
wtforms
